refactor(api): move serializers to prescription bounded context
