+++ api/src/prescription/learner-list/application/learner-list-controller.js
@@ -1,5 +1,5 @@
 import { usecases } from '../domain/usecases/index.js';
-import * as organizationParticipantsSerializer from '../../../../lib/infrastructure/serializers/jsonapi/organization/organization-participants-serializer.js';
+import * as organizationParticipantsSerializer from '../infrastructure/serializers/jsonapi/organization-participants-serializer.js';
 import { mapCertificabilityByLabel } from './helpers.js';
 import * as queryParamsUtils from '../../../../lib/infrastructure/utils/query-params-utils.js';
