+++ api/src/prescription/learner-list/application/sco-learner-list-controller.js
@@ -1,5 +1,5 @@
 import { usecases } from '../domain/usecases/index.js';
-import * as scoOrganizationParticipantsSerializer from '../../../../lib/infrastructure/serializers/jsonapi/organization/sco-organization-participants-serializer.js';
+import * as scoOrganizationParticipantsSerializer from '../infrastructure/serializers/jsonapi/sco-organization-participants-serializer.js';
 import { mapCertificabilityByLabel } from './helpers.js';
 import * as queryParamsUtils from '../../../../lib/infrastructure/utils/query-params-utils.js';
