+++ api/src/prescription/learner-list/infrastructure/serializers/jsonapi/organization-participants-serializer.js
