+++ api/src/prescription/learner-list/infrastructure/serializers/jsonapi/sco-organization-participants-serializer.js
