+++ api/src/prescription/learner-list/infrastructure/serializers/jsonapi/sup-organization-participants-serializer.js
